GRHB-408: * rename my-stydents.tsx to course-students.tsx
