+++ mobile/src/components/course/components/components.ts
@@ -1,5 +1,5 @@
 export { ChooseMentor } from './choose-mentor/choose-mentor';
 export { CourseCategory } from './course-category/course-category';
 export { CourseDescription } from './course-description/course-description';
-export { CourseStudents } from './course-students/my-stydents';
+export { CourseStudents } from './course-students/course-students';
 export { EditCourseCategory } from './edit-course-category/edit-course-category';
